feat: Add ASCII art files for bottles, swords, skulls, treasure, ships, and pirates
The commit adds ASCII art files for bottles, swords, skulls, treasure, ships, and pirates. These files are added to the `art` directory. The ASCII art files are in different sizes and depict various objects related to the theme of a ghost ship. This commit is part of the initial commit to the repository.